rework - global ImageManager for uicon json index
